+++ ogc-example/src/main/scala/geotrellis/server/ogc/Main.scala
@@ -118,7 +118,8 @@ object Main extends CommandApp(
               WmsModel(
                 svc.serviceMetadata,
                 svc.parentLayerMeta,
-                svc.layerSources(simpleSources)
+                svc.layerSources(simpleSources),
+                FocalParameters.extendedParametersBinding
               )
             }
             _ <- Stream.eval(IO(logOptState(
@@ -141,7 +142,8 @@ object Main extends CommandApp(
             wcsModel = conf.wcs.map { svc =>
               WcsModel(
                 svc.serviceMetadata,
-                svc.layerSources(simpleSources)
+                svc.layerSources(simpleSources),
+                FocalParameters.extendedParametersBinding
               )
             }
             ogcService = new OgcService(wmsModel, wcsModel, wmtsModel, new URL(publicUrl))